% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanDeviceFunctions.R
\name{closeDeviceAll}
\alias{closeDeviceAll}
\title{closeDeviceAll}
\usage{
closeDeviceAll()
}
\description{
Close all devices for printing a report
}
\details{
Notes:
The  MSP's are in units of MeV pr. g/cm2.
The I is in eV
The density (rho) is in g/cm3
}
