% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanDeviceFunctions.R
\name{demo.clanDevice}
\alias{demo.clanDevice}
\title{demo.clanDevice}
\usage{
demo.clanDevice()
}
\description{
Demonstration of how to use the clanDevice package
}
\details{
Notes:
Make a pdf-repost with three plote
}
