% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanDeviceFunctions.R
\name{openDevice}
\alias{openDevice}
\title{openDevice}
\usage{
openDevice(
  grand.plotfilename = "junk",
  type = "pdf",
  png.fac = 1.5,
  png.width = png.fac * 20,
  png.height = png.fac * 13,
  png.res = 600,
  png.bg = "white",
  png.family = "arial",
  png.type = c("windows", "cairo", "cairo-png")[2],
  png.restoreConsole = TRUE,
  pdf.inch.fac = 2.54/1.2,
  pdf.width = 29.7/pdf.inch.fac,
  pdf.height = 21/pdf.inch.fac,
  pdf.pointsize = 19,
  pdf.family = "Courier"
)
}
\arguments{
\item{grand.plotfilename}{= "junk" (base file name for output)}

\item{type}{="png","png.linux", "pdf", "ps", or "none"}

\item{png.fac}{= 1.5 (scaling factor for png)}

\item{png.width}{= ng.fac*20 (width)}

\item{png.height}{= png.fac*13 (height)}

\item{png.res}{= 600 (resolution)}

\item{png.bg}{= "white" (lightgray or pink are also good background colors)}

\item{png.family}{= "arial" (font)}

\item{png.type}{= c("windows", "cairo", "cairo-png")[2] on Windows}

\item{png.restoreConsole}{= TRUE on Windows}

\item{pdf.inch.fac}{= 2.54 / 1.2 (scaling factor for pdf)}

\item{pdf.width}{= 9.7/pdf.inch.fac}

\item{pdf.height}{= 21/pdf.inch.fac}

\item{pdf.pointsize}{= 19}

\item{pdf.family}{= "Courier" (font)}
}
\description{
Open device for printing a report
}
\details{
Notes:
The png.linux is without the restoreConsole parameter (only available on windows).
}
