% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanDeviceFunctions.R
\name{openDevice}
\alias{openDevice}
\title{openDevice}
\usage{
openDevice(
  grand.plotfilename = "junk",
  type = "pdf",
  png.fac = 1.5,
  png.width = png.fac * 20,
  png.height = png.fac * 13,
  png.res = 600,
  png.bg = "white",
  png.family = "arial",
  png.type = c("windows", "cairo", "cairo-png")[2],
  pdf.inch.fac = 2.54/1.2,
  pdf.width = 29.7/pdf.inch.fac,
  pdf.height = 21/pdf.inch.fac,
  pdf.pointsize = 19,
  pdf.family = "Courier"
)
}
\arguments{
\item{grand.plotfilename}{= "junk"}

\item{type}{="png", "pdf", "ps", or "none"}

\item{png.fac}{= 1}

\item{png.width}{= date frame with data (df.NIST.estar.SP)}

\item{png.height}{= rule for approx: 1=no extrapolation, 2=use nearest value for extrapolation}
}
\description{
Open device for printing a report
}
\details{
Notes:
The  MSP's are in units of MeV pr. g/cm2.
The I is in eV
The density (rho) is in g/cm3
}
